drop search index if it exists

+++ bookshelf/bin/create_indexes.rb
@@ -13,7 +13,9 @@ options = Couchbase::Management::QueryIndexManager::CreatePrimaryIndexOptions.ne
 options.ignore_if_exists = true
 cluster.query_indexes.create_primary_index(config["bucket"], options)
 
-cluster.search_indexes.drop_index("books-index")
+if cluster.search_indexes.get_all_indexes.any? { |idx| idx.name == "books-index" }
+  cluster.search_indexes.drop_index("books-index")
+end
 index = Couchbase::Management::SearchIndex.new
 index.type = "fulltext-index"
 index.name = "books-index"
